# improve String#to_proc perf