Builds on https://github.com/opal/opal/commit/6ce0619e0e1acdb2ba8ca406263bd8a2a329a91f

Fixes #1659 

reaching 1.06x vs the inlined block

cc @mojavelinux @Mogztter