@elia 2018-10-02
IPS benchmark results (see https://github.com/opal/opal/pull/1888/commits/3868d5e0aaeb27c92d195fe7d5e55d249f1521d9)

```
BEFORE

opal:elia/string-to-proc-perf ⤑ bundle exec rake bench:ips FILE=benchmark-ips/bm_symbol_to_proc.rb                                     ~/C/opal
=== Files: benchmark-ips/bm_symbol_to_proc.rb
bundle exec opal --dynamic-require ignore --missing-require ignore -ropal/platform -gbenchmark-ips -rbenchmark/ips -A benchmark-ips/bm_symbol_to_proc.rb
Warming up --------------------------------------
           map block     1.090k i/100ms
          map symbol   869.000  i/100ms
Calculating -------------------------------------
           map block     11.302k (± 1.0%) i/s -     56.680k in   5.015370s
          map symbol      8.878k (± 2.3%) i/s -     45.188k in   5.092622s

Comparison:
           map block:    11302.5 i/s
          map symbol:     8878.1 i/s - 1.27x  slower



AFTER

opal:master ⤑ bundle exec rake bench:ips FILE=benchmark-ips/bm_symbol_to_proc.rb                                                       ~/C/opal
=== Files: benchmark-ips/bm_symbol_to_proc.rb
bundle exec opal --dynamic-require ignore --missing-require ignore -ropal/platform -gbenchmark-ips -rbenchmark/ips -A benchmark-ips/bm_symbol_to_proc.rb
Warming up --------------------------------------
           map block   980.000  i/100ms
          map symbol   977.000  i/100ms
Calculating -------------------------------------
           map block     10.430k (± 1.1%) i/s -     52.920k in   5.074384s
          map symbol      9.804k (± 2.6%) i/s -     49.827k in   5.086232s

Comparison:
           map block:    10430.1 i/s
          map symbol:     9803.7 i/s - 1.06x  slower
```