@mojavelinux 2018-10-03
Nice work. Every day Opal gets faster is a good day for me ;)